+++ albums/mock_cache_test.go
@@ -8,6 +8,7 @@ import (
 type MockedCache struct {
 	GetAlbumFn            func(ctx context.Context, title string) (Album, error)
 	PutAlbumFn            func(ctx context.Context, album Album) error
+	PutManyAlbumsFn       func(ctx context.Context, albums []Album) error
 	InvalidateAlbumFn     func(ctx context.Context, title string) error
 	InvalidateAllAlbumsFn func(ctx context.Context) error
 }
@@ -20,6 +21,11 @@ func (c MockedCache) GetAlbum(ctx context.Context, title string) (Album, error) 
 // PutAlbum invokes the mock implementation.
 func (c MockedCache) PutAlbum(ctx context.Context, album Album) error {
 	return c.PutAlbumFn(ctx, album)
+}
+
+// PutManyAlbums invokes the mock implementation.
+func (c MockedCache) PutManyAlbums(ctx context.Context, albums []Album) error {
+	return c.PutManyAlbumsFn(ctx, albums)
 }
 
 // InvalidateAlbum invokes the mock implementation.